+++ src/main/java/racingcar/domain/model/Car.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain.model;
 
 import java.util.Arrays;
 import java.util.HashMap;
