+++ src/main/java/racingcar/service/Game.java
@@ -1,0 +1,7 @@
+package racingcar.domain;
+
+public class Game {
+    public void playGame() {
+
+    }
+}
