+++ src/main/java/racingcar/ui/output/OutputProcessor.java
@@ -1,11 +1,15 @@
-package racingcar.domain;
+package racingcar.output;
 
-public class Output {
+public class OutputProcessor {
     public void printInput(String input) {
         System.out.println(input);
     }
 
     public void printResult(String carName, int forwardNumber) {
         System.out.println(carName+" "+": "+"-".repeat(forwardNumber));
+    }
+
+    public void printStartMessage() {
+
     }
 }
